// src/interfaces/ComponentFunctionalityInterface.ts
export interface ComponentFunctionality {
    bootstrapTooltip: () => void;
    bulmaAlert: () => void;
    bulmaAccordion: () => void;
    bulmaModal: () => void;
    bulmaTrigger: () => void;
    foundationTrigger: () => void;
    materializeAccordion: () => void;
    materializeDropdown: () => void;
    materializeModal: () => void;
    materializeTooltip: () => void;
    pureDropdown: () => void;
    semanticAccordion: () => void;
    semanticAlert: () => void;
    semanticDropdown: () => void;
    semanticModal: () => void;
    spectreModal: () => void;
  }